# Changed LSP completion resolve to not cache data on items.