--- a/src/Features/LanguageServer/Protocol/IsExternalInit.cs
+++ b/src/Features/LanguageServer/Protocol/IsExternalInit.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Runtime.CompilerServices
+{
+    // Used to compile against C# 9 in a netstandard2.0
+    internal class IsExternalInit
+    {
+    }
+}