… src/Features/LanguageServer/ProtocolUnitTests/Completion/CompletionTests.cs